PetrasGreenClock/requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
MouseInfo==0.1.3
numpy==2.3.1
packaging==25.0
pillow==11.2.1
pip==23.1.2
pluggy==1.6.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
pytweening==1.2.0
pytz==2025.2
pywin32==310
requests==2.32.4
screeninfo==0.8.1
setuptools==65.5.0
six==1.17.0
urllib3==2.5.0
python-dotenv
